chore: resolve conflict

+++ packages/fx-core/src/core/middleware/projectMigrator.ts
@@ -19,12 +19,13 @@ import {
 import {
   CoreHookContext,
   deserializeDict,
-  NoProjectOpenedError,
   serializeDict,
   SolutionConfigError,
   ProjectSettingError,
   environmentManager,
+  SPFxConfigError,
 } from "../..";
+import { UpgradeCanceledError } from "../error";
 import { LocalSettingsProvider } from "../../common/localSettingsProvider";
 import { Middleware, NextFunction } from "@feathersjs/hooks/lib";
 import fs from "fs-extra";
@@ -51,7 +52,7 @@ import { loadSolutionContext } from "./envInfoLoader";
 import { ResourcePlugins } from "../../common/constants";
 import { getActivatedResourcePlugins } from "../../plugins/solution/fx-solution/ResourcePluginContainer";
 import { LocalDebugConfigKeys } from "../../plugins/resource/localdebug/constants";
-import { MANIFEST_LOCAL } from "../../plugins/resource/appstudio/constants";
+import { MANIFEST_LOCAL, MANIFEST_TEMPLATE } from "../../plugins/resource/appstudio/constants";
 import {
   Component,
   ProjectMigratorGuideStatus,
@@ -60,16 +61,21 @@ import {
   TelemetryEvent,
   TelemetryProperty,
 } from "../../common/telemetry";
+import { PlaceHolders } from "../../plugins/resource/spfx/utils/constants";
+import { Utils as SPFxUtils } from "../../plugins/resource/spfx/utils/utils";
 
 const programmingLanguage = "programmingLanguage";
 const defaultFunctionName = "defaultFunctionName";
 const learnMoreText = "Learn More";
 const reloadText = "Reload";
+const upgradeButton = "Upgrade";
 const solutionName = "solution";
 const subscriptionId = "subscriptionId";
 const resourceGroupName = "resourceGroupName";
 const migrationGuideUrl = "https://aka.ms/teamsfx-migration-guide";
 const parameterFileNameTemplate = "azure.parameters.@envName.json";
+const learnMoreLink =
+  "https://github.com/OfficeDev/TeamsFx/wiki/Upgrade-project-to-use-latest-Toolkit-features";
 let updateNotificationFlag = false;
 let fromReloadFlag = false;
 
@@ -115,13 +121,20 @@ export const ProjectMigratorMW: Middleware = async (ctx: CoreHookContext, next: 
       "warn",
       getStrings().solution.MigrationToArmAndMultiEnvMessage,
       true,
-      "OK"
+      upgradeButton
     );
     const answer = res?.isOk() ? res.value : undefined;
-    if (!answer || answer != "OK") {
+    if (!answer || answer != upgradeButton) {
       sendTelemetryEvent(Component.core, TelemetryEvent.ProjectMigratorNotification, {
         [TelemetryProperty.Status]: ProjectMigratorStatus.Cancel,
       });
+      ctx.result = err(UpgradeCanceledError());
+      core.tools.logProvider.warning(
+        `[core] Upgrade canceled. If you don't want to upgrade your project, please install another version of Teams Toolkit(version <= 2.7.0). Read this wiki(${learnMoreLink}) to learn how to downgrade Teams Toolkit.`
+      );
+      core.tools.logProvider.warning(
+        `[core] You can also reload VSCode or run command (Teams: Check project upgrade) to trigger upgrade check.`
+      );
       return;
     }
     sendTelemetryEvent(Component.core, TelemetryEvent.ProjectMigratorNotification, {
@@ -129,12 +142,15 @@ export const ProjectMigratorMW: Middleware = async (ctx: CoreHookContext, next: 
     });
 
     await migrateToArmAndMultiEnv(ctx);
+    core.tools.logProvider.warning(
+      `[core] Upgrade success! All old files in .fx and appPackage folder have been backed up to the .backup folder and you can delete it. Read this wiki(${learnMoreLink}) if you want to restore your project or learn more about this upgrade.`
+    );
   } else if ((await needUpdateTeamsToolkitVersion(ctx)) && !updateNotificationFlag) {
     // TODO: delete before Arm && Multi-env version released
     // only for arm && multi-env project with unreleased teams toolkit version
     updateNotificationFlag = true;
     const core = ctx.self as FxCore;
-    await core.tools.ui.showMessage(
+    core.tools.ui.showMessage(
       "info",
       getStrings().solution.NeedToUpdateTeamsToolkitVersionMessage,
       false,
@@ -283,7 +299,7 @@ async function migrateMultiEnv(projectPath: string): Promise<void> {
 
   // appPackage
   await fs.copy(path.join(projectPath, AppPackageFolderName), templateAppPackage);
-  const targetManifestFile = path.join(templateAppPackage, "manifest.template.json");
+  const targetManifestFile = path.join(templateAppPackage, MANIFEST_TEMPLATE);
   await fs.rename(path.join(templateAppPackage, "manifest.source.json"), targetManifestFile);
 
   // update manifest to mustache template
@@ -291,43 +307,57 @@ async function migrateMultiEnv(projectPath: string): Promise<void> {
   manifestString = manifestString.replace(new RegExp("{version}", "g"), "1.0.0");
   manifestString = manifestString.replace(
     new RegExp("{baseUrl}", "g"),
-    "{{{profile.fx-resource-frontend-hosting.endpoint}}}"
+    "{{{state.fx-resource-frontend-hosting.endpoint}}}"
   );
   manifestString = manifestString.replace(
     new RegExp("{appClientId}", "g"),
-    "{{profile.fx-resource-aad-app-for-teams.clientId}}"
+    "{{state.fx-resource-aad-app-for-teams.clientId}}"
   );
   manifestString = manifestString.replace(
     new RegExp("{webApplicationInfoResource}", "g"),
-    "{{{profile.fx-resource-aad-app-for-teams.applicationIdUris}}}"
+    "{{{state.fx-resource-aad-app-for-teams.applicationIdUris}}}"
   );
   manifestString = manifestString.replace(
     new RegExp("{botId}", "g"),
-    "{{profile.fx-resource-bot.botId}}"
+    "{{state.fx-resource-bot.botId}}"
   );
   const manifest: TeamsAppManifest = JSON.parse(manifestString);
   manifest.name.short = "{{config.manifest.appName.short}}";
   manifest.name.full = "{{config.manifest.appName.full}}";
-  manifest.id = "{{profile.fx-resource-appstudio.teamsAppId}}";
+  manifest.id = "{{state.fx-resource-appstudio.teamsAppId}}";
   await fs.writeFile(targetManifestFile, JSON.stringify(manifest, null, 4));
   await moveIconsToResourceFolder(templateAppPackage);
 
-  if (!isSPFx) {
-    const localManifest: TeamsAppManifest = createLocalManifest(
-      appName,
-      hasFrontend,
-      hasBotCapability,
-      hasMessageExtensionCapability
-    );
-    const localManifestFile = path.join(templateAppPackage, MANIFEST_LOCAL);
-    await fs.writeFile(localManifestFile, JSON.stringify(localManifest, null, 4));
+  // generate manifest.local.template.json
+  const localManifest: TeamsAppManifest = await createLocalManifest(
+    appName,
+    hasFrontend,
+    hasBotCapability,
+    hasMessageExtensionCapability,
+    isSPFx
+  );
+  const localManifestFile = path.join(templateAppPackage, MANIFEST_LOCAL);
+  await fs.writeFile(localManifestFile, JSON.stringify(localManifest, null, 4));
+
+  if (isSPFx) {
+    const replaceMap: Map<string, string> = new Map();
+    const packageSolutionFile = `${projectPath}/SPFx/config/package-solution.json`;
+    if (!(await fs.pathExists(packageSolutionFile))) {
+      throw SPFxConfigError(packageSolutionFile);
+    }
+    const solutionConfig = await fs.readJson(packageSolutionFile);
+    replaceMap.set(PlaceHolders.componentId, solutionConfig.solution.id);
+    replaceMap.set(PlaceHolders.componentNameUnescaped, manifest.packageName!);
+    await SPFxUtils.configure(localManifestFile, replaceMap);
   }
 
   if (hasProvision) {
     const devState = path.join(fxState, "state.dev.json");
     const devUserData = path.join(fxState, "dev.userdata");
     await fs.copy(path.join(fx, "new.env.default.json"), devState);
-    await fs.copy(path.join(fx, "default.userdata"), devUserData);
+    if (await fs.pathExists(path.join(fx, "default.userdata"))) {
+      await fs.copy(path.join(fx, "default.userdata"), devUserData);
+    }
     await removeExpiredFields(devState, devUserData);
   }
 }
@@ -335,7 +365,7 @@ async function migrateMultiEnv(projectPath: string): Promise<void> {
 async function moveIconsToResourceFolder(templateAppPackage: string): Promise<void> {
   // see AppStudioPluginImpl.buildTeamsAppPackage()
   const manifest: TeamsAppManifest = await readJson(
-    path.join(templateAppPackage, "manifest.template.json")
+    path.join(templateAppPackage, MANIFEST_TEMPLATE)
   );
   const hasColorIcon = manifest.icons.color && !manifest.icons.color.startsWith("https://");
   const hasOutlineIcon = manifest.icons.outline && !manifest.icons.outline.startsWith("https://");
@@ -358,19 +388,18 @@ async function moveIconsToResourceFolder(templateAppPackage: string): Promise<vo
   manifest.icons.color = `resources/${manifest.icons.color}`;
   manifest.icons.outline = `resources/${manifest.icons.outline}`;
   await fs.writeFile(
-    path.join(templateAppPackage, "manifest.template.json"),
+    path.join(templateAppPackage, MANIFEST_TEMPLATE),
     JSON.stringify(manifest, null, 4)
   );
 }
 
-async function removeExpiredFields(devProfile: string, devUserData: string): Promise<void> {
-  const profileData = await readJson(devProfile);
+async function removeExpiredFields(devState: string, devUserData: string): Promise<void> {
+  const stateData = await readJson(devState);
   const secrets: Record<string, string> = deserializeDict(await fs.readFile(devUserData, "UTF-8"));
 
-  profileData[PluginNames.APPST]["teamsAppId"] =
-    profileData[PluginNames.SOLUTION]["remoteTeamsAppId"];
+  stateData[PluginNames.APPST]["teamsAppId"] = stateData[PluginNames.SOLUTION]["remoteTeamsAppId"];
 
-  const expiredProfileKeys: [string, string][] = [
+  const expiredStateKeys: [string, string][] = [
     [PluginNames.LDEBUG, ""],
     [PluginNames.SOLUTION, programmingLanguage],
     [PluginNames.SOLUTION, defaultFunctionName],
@@ -385,12 +414,12 @@ async function removeExpiredFields(devProfile: string, devUserData: string): Pro
     [PluginNames.SA, "filePath"],
     [PluginNames.SA, "environmentVariableParams"],
   ];
-  for (const [k, v] of expiredProfileKeys) {
-    if (profileData[k]) {
+  for (const [k, v] of expiredStateKeys) {
+    if (stateData[k]) {
       if (!v) {
-        delete profileData[k];
-      } else if (profileData[k][v]) {
-        delete profileData[k][v];
+        delete stateData[k];
+      } else if (stateData[k][v]) {
+        delete stateData[k][v];
       }
     }
   }
@@ -400,7 +429,7 @@ async function removeExpiredFields(devProfile: string, devUserData: string): Pro
   }
   deleteUserDataKey(secrets, `${PluginNames.AAD}.local_clientSecret`);
 
-  await fs.writeFile(devProfile, JSON.stringify(profileData, null, 4), { encoding: "UTF-8" });
+  await fs.writeFile(devState, JSON.stringify(stateData, null, 4), { encoding: "UTF-8" });
   await fs.writeFile(devUserData, serializeDict(secrets), { encoding: "UTF-8" });
 }
 
@@ -452,10 +481,22 @@ async function getMultiEnvFolders(projectPath: string): Promise<any> {
   return { fx, fxConfig, templateAppPackage, fxState };
 }
 
+async function getBackupFolder(projectPath: string): Promise<string> {
+  const backupName = ".backup";
+  const backup = path.join(projectPath, backupName);
+  if (!(await fs.pathExists(backup))) {
+    return backup;
+  }
+  // avoid conflict(rarely)
+  return path.join(projectPath, `.teamsfx${backupName}`);
+}
 async function backup(projectPath: string): Promise<void> {
   const fx = path.join(projectPath, `.${ConfigFolderName}`);
-  const backup = path.join(fx, "migrationbackup");
-  await fs.ensureDir(backup);
+  const backup = await getBackupFolder(projectPath);
+  const backupFx = path.join(backup, `.${ConfigFolderName}`);
+  const backupAppPackage = path.join(backup, AppPackageFolderName);
+  await fs.ensureDir(backupFx);
+  await fs.ensureDir(backupAppPackage);
   const fxFiles = [
     "env.default.json",
     "default.userdata",
@@ -466,17 +507,14 @@ async function backup(projectPath: string): Promise<void> {
 
   for (const file of fxFiles) {
     if (await fs.pathExists(path.join(fx, file))) {
-      await fs.copy(path.join(fx, file), path.join(backup, file));
+      await fs.copy(path.join(fx, file), path.join(backupFx, file));
     }
   }
   if (await fs.pathExists(path.join(projectPath, AppPackageFolderName))) {
-    await fs.copy(
-      path.join(projectPath, AppPackageFolderName),
-      path.join(backup, AppPackageFolderName)
-    );
+    await fs.copy(path.join(projectPath, AppPackageFolderName), backupAppPackage);
   } else if (await fs.pathExists(path.join(fx, AppPackageFolderName))) {
     // version <= 2.4.1
-    await fs.copy(path.join(fx, AppPackageFolderName), path.join(backup, AppPackageFolderName));
+    await fs.copy(path.join(fx, AppPackageFolderName), backupAppPackage);
   }
 }
